+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Enh #319: Support `boolean` type (@Tigrov)
 - Enh #318, #320: Use `DbArrayHelper::arrange()` instead of `DbArrayHelper::index()` method (@Tigrov)
 - New #316: Realize `Schema::loadResultColumn()` method (@Tigrov)
+- New #323: Use `DateTimeColumn` class for datetime column types (@Tigrov)
 
 ## 1.3.0 March 21, 2024